Update make script
The existing code was trying `echo "\n"`, which doesn't work. It outputs `\n` instead of an actual newline. They need to be wrapped with ANSI C quotes (`$'\n'`), which (according to the manual):

```
Words of the form $'string' are treated specially. The word expands to string, with backslash-escaped characters replaced as specified by the ANSI C standard.
```

Add indentation
Make it so the file sets the proper error code
Replace hard-coded script name with `$0`